[#388] Remove CPU serial number.
For the migration to Bookworm many users will need to change the hardware.
The license should therefore not be coupled to the CPU serial number anymore.
(On amd64 it was a dummy value, anyway.)
The default registration type is now 0 (independent of CPU serial).
Migrated licenses of type 1 are validated without checking the CPU serial.